Add files via upload
This dataset contains detailed information on road crashes that occurred from 2012 to 2023. Each record includes data on the accident number, date, time, type, and a descriptive code of the accident. Additional details such as the day of the week, light conditions, road geometry, severity of the crash, speed zone, sensors, and road management authority are also provided.